requirements.txt
Pillow==2.9.0
numpy==1.10.1
